ragSystem/requirementsPython.txt
langchain

langchain-openai
openai

langchain-anthropic
langchain-google-genai
google-generativeai
langchain-huggingface
langchain_chroma
transformers
huggingface-hub
python-dotenv
numpy
scikit-learn

python-multipart
ipykernel
chromadb

fastapi 
uvicorn

PyPDF2